requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
pydantic
python-multipart
cryptography
rapidocr_onnxruntime
passlib[bcrypt]
python-jose[cryptography]